Adversarial review batch 3: naming and declaration fixes
Naming-lens findings from the adversarial subagent sweep, all verified
MATCH + CHECK-GREEN:
- ChkCard.c: the memory-card TU note pointed at the wrong address
  (0x8001b4bc is another region); the function lives at 0x80056df8.
- LoadSI.c: fmt_card_name's comment claimed "%s\%d\%s"; the actual
  string bytes (and SaveSI's copy) are "%s%d_%s".
- SelectCameraOwnerOption.c: header size corrected 0x158 -> 0x110
  (tdis: 272 bytes).
- game_types.h: dropped the orphaned commented-out AdtSelect prototype.
- AdtSelect.c: renamed the mode parameter from the invented `selection`
  to `mode`, matching every caller-side extern (no official name in the
  demo symbols).
- game_over_screen_.c: FadeOutDirect extern's stray `s32 b` -> `u8 b`
  to agree with the definition (measured byte-free).
- set_model_hide_.c: raw | 1 / & ~1 loop bodies now spell
  MODEL_ATTR_HIDDEN.
- item.h + MoveKorogari.c: retail-only rolling-item state 5 named
  KORO_OUT (invented; fires when CGetLevel says the item left the map).
- CreateStage.c: STAGE_LAYOUT_NUMBER's volatile was unnecessary
  (measured byte-free without it).
- remap_buttons_.c: ControlScheme MUST stay s16 here (u16 flips lh ->
  lhu); noted. ActivateHumans's s32 PacketUsed likewise stays (image
  drift when changed); both TU-local extern divergences are real.
- Think3callaid.c: cross-ref note that AIDHumanType's flat shape here
  vs [][2] in think_alarm_reaction_.c is the required pointer idiom.
- Seven files: collapsed doubled string comments to the quoted form;
  DoInfoViewProc's two disagreeing comments merged; retitled
  debug_menu_stage_option's fmt_layout_no comment (it is a debug stats
  dump, not a score message).

+++ src/main.exe/game_over_screen_.c
@@ -37,7 +37,7 @@ extern char path_demo[];                  /* K:\\WORK\\CDIMAGE\\DEMO\\ */
 extern char *GOV_RESOURCE_PREFIX_PTRS[];
 extern char *GOV_ARCHIVE_PTRS[];
 
-extern void FadeOutDirect(s16 time, s16 attrib, u8 r, u8 g, s32 b);
+extern void FadeOutDirect(s16 time, s16 attrib, u8 r, u8 g, u8 b);
 extern void clear_screen_(void);
 extern s32 VSync(s32 mode);
 extern Sprite3D *SetupSprite(Sprite3D *orgsprt, GsIMAGE *image);
